Split operating-specific functionality into separate files
